feat(ui): code pallette widget
