+++ src/feature/About/styles.module.scss
@@ -39,10 +39,12 @@
   display: flex;
   flex-direction: column;
   overflow: hidden;
+  justify-content: space-between;
+
   gap: 20px;
-  border: 1px solid;
   padding: 24px;
   border-radius: var(--default-border-radius);
+  background: linear-gradient(to top left, #3a3a3a 0%, #242424 50%, #3a3a3a 100%);
 
   &:nth-child(1) {
     grid-area: first;
@@ -61,18 +63,15 @@
   }
 }
 
+.headerContainer {
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+}
+
 .locationText {
   text-wrap: balance;
   color: var(--secondary-text-color);
-}
-
-.globeWidget {
-  background: linear-gradient(
-    to top left,
-    #3a3a3a 0%,
-    /* rgb(58, 58, 58) */ #242424 50%,
-    /* rgb(36, 36, 36) */ #3a3a3a 100% /* rgb(58, 58, 58) */
-  );
 }
 
 .globWrapper {
